Renamed generate_db -> generate_test_db, removed unnecessary character in twitter_search
