+++ foodbornenyc/sources/twitter_search.py
@@ -99,7 +99,7 @@ def query_twitter(how_long=0):
         if not tweets: # if we dont get anything back, sleep and try again
             time.sleep(request_wait)
             continue
-        logger.info("%i Total unique tweets in %i:%i:%i time", len(id_set),\
+        logger.info("%i Total unique tweets in %i:%i:%i time", len(id_set),
                     *sec_to_hms(time.time()-start))
         new_tweets = [ t for t in tweets if t['id_str'] not in id_set ]
         new_Tweets = tweets_to_Tweets(new_tweets, fields)
